# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This service provides sends aggregate and tracker data to DHIS2 from third-party systems",
        "title": "DHIS2 Gateway Service",
        "contact": {
            "name": "API Support",
            "url": "http://www.hispuganda.org",
            "email": "ssekiwere@hispuganda.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "host": "dhis2gw.hispuganda.org",
    "basePath": "/api/v2",
    "paths": {
...
        "/logs": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Returns a paginated list of job logs with optional filters like status, task ID, job ID, and submission date range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get job logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by task id",
                        "name": "task_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by job id",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by exact submitted_at (RFC3339)",
                        "name": "submitted_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Submitted after (RFC3339)",
                        "name": "submitted_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Submitted before (RFC3339)",
                        "name": "submitted_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.JobLogPaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server-side error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/logs/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get a specific job log entry by its database ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get job log by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job log entry",
                        "schema": {
                            "$ref": "#/definitions/joblog.JobLogSwagger"
                        }
                    },
                    "404": {
                        "description": "Log not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server-side error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.JobLogPaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/joblog.JobLogSwagger"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "joblog.JobLogSwagger": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "last_attempt_at": {
                    "type": "string",
                    "example": "2024-06-24T09:00:00Z"
                },
                "payload": {
                    "type": "object"
                },
                "response": {
                    "type": "string",
                    "example": "OK"
                },
                "retry_count": {
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "type": "string",
                    "example": "SUCCESS"
                },
                "submitted_at": {
                    "type": "string",
                    "example": "2024-06-24T08:00:00Z"
                },
                "task_id": {
                    "type": "string",
                    "example": "abc-123"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {},
                "error": {
                    "type": "string",
                    "example": "Invalid JSON"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "security": [
        {
            "basicAuth": []
        }
    ]
}